+++ src/main/java/com/example/Server/global/oauth/common/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.example.Server.oauth.common.config;
+package com.example.Server.global.oauth.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
