+++ src/main/java/com/example/Server/global/oauth/member/domain/Role.java
@@ -1,0 +1,6 @@
+package com.example.Server.global.oauth.member.domain;
+
+public enum Role {
+    USER,
+    ADMIN
+}
