+++ src/main/java/com/example/Server/global/oauth/member/dto/AccessTokenDto.java
@@ -1,4 +1,4 @@
-package com.example.Server.oauth.member.dto;
+package com.example.Server.global.oauth.member.dto;
 
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
