+++ src/main/java/com/example/Server/global/oauth/member/service/MemberService.java
@@ -1,6 +1,6 @@
-package com.example.Server.oauth.member.service;
+package com.example.Server.global.oauth.member.service;
 
-import com.example.Server.oauth.member.domain.Member;
+import com.example.Server.global.oauth.member.domain.Member;
 
 public interface MemberService {
