--- src/main/java/com/example/Server/oauth/member/domain/Role.java
@@ -1,6 +1,0 @@
-package com.example.Server.oauth.member.domain;
-
-public enum Role {
-    USER,
-    ADMIN
-}
